+++ src/main/java/edu/insightr/gildedrose/Model/Backstage_Passes.java
@@ -1,4 +1,4 @@
-package edu.insightr.gildedrose;
+package edu.insightr.gildedrose.Model;
 
 public class Backstage_Passes extends Item{
