+++ src/main/java/edu/insightr/gildedrose/Model/Dexterity_Vest.java
@@ -1,4 +1,4 @@
-package edu.insightr.gildedrose;
+package edu.insightr.gildedrose.Model;
 
 public class Dexterity_Vest extends Item{
