+++ src/main/java/edu/insightr/gildedrose/Model/IVisitable.java
@@ -1,4 +1,4 @@
-package edu.insightr.gildedrose;
+package edu.insightr.gildedrose.Model;
 
 public interface IVisitable {
     void accept(IVisitor visitor);
