+++ src/main/java/edu/insightr/gildedrose/Model/IVisitor.java
@@ -1,4 +1,4 @@
-package edu.insightr.gildedrose;
+package edu.insightr.gildedrose.Model;
 
 public interface IVisitor {
     void visit(IVisitable o);
