+++ src/main/java/edu/insightr/gildedrose/Model/Sulfuras.java
@@ -1,4 +1,4 @@
-package edu.insightr.gildedrose;
+package edu.insightr.gildedrose.Model;
 
 public class Sulfuras extends Item{
